+++ cardano-cli/src/Cardano/CLI/Shelley/Run/Transaction.hs
@@ -266,11 +266,11 @@ renderFeature TxFeatureReturnCollateral     = "Return collateral"
 runTransactionCmd :: TransactionCmd -> ExceptT ShelleyTxCmdError IO ()
 runTransactionCmd cmd =
   case cmd of
-    TxBuild era consensusModeParams nid mScriptValidity mOverrideWits txins readOnlyRefIns
+    TxBuild mNodeSocketPath era consensusModeParams nid mScriptValidity mOverrideWits txins readOnlyRefIns
             reqSigners txinsc mReturnColl mTotCollateral txouts changeAddr mValue mLowBound
             mUpperBound certs wdrls metadataSchema scriptFiles metadataFiles mProtocolParamsFile
             mUpProp outputOptions -> do
-      runTxBuildCmd era consensusModeParams nid mScriptValidity mOverrideWits txins readOnlyRefIns
+      runTxBuildCmd mNodeSocketPath era consensusModeParams nid mScriptValidity mOverrideWits txins readOnlyRefIns
             reqSigners txinsc mReturnColl mTotCollateral txouts changeAddr mValue mLowBound
             mUpperBound certs wdrls metadataSchema scriptFiles metadataFiles mProtocolParamsFile mUpProp outputOptions
     TxBuildRaw era mScriptValidity txins readOnlyRefIns txinsc mReturnColl
@@ -281,8 +281,8 @@ runTransactionCmd cmd =
                metadataSchema scriptFiles metadataFiles mProtocolParamsFile mUpProp out
     TxSign txinfile skfiles network txoutfile ->
       runTxSign txinfile skfiles network txoutfile
-    TxSubmit anyConsensusModeParams network txFp ->
-      runTxSubmit anyConsensusModeParams network txFp
+    TxSubmit mNodeSocketPath anyConsensusModeParams network txFp ->
+      runTxSubmit mNodeSocketPath anyConsensusModeParams network txFp
     TxCalculateMinFee txbody mnw pParamsFile nInputs nOutputs nShelleyKeyWitnesses nByronKeyWitnesses ->
       runTxCalculateMinFee txbody mnw pParamsFile nInputs nOutputs nShelleyKeyWitnesses nByronKeyWitnesses
     TxCalculateMinRequiredUTxO era pParamsFile txOuts -> runTxCalculateMinRequiredUTxO era pParamsFile txOuts
@@ -300,7 +300,8 @@ runTransactionCmd cmd =
 --
 
 runTxBuildCmd
-  :: AnyCardanoEra
+  :: Maybe SocketPath
+  -> AnyCardanoEra
   -> AnyConsensusModeParams
   -> NetworkId
   -> Maybe ScriptValidity
@@ -326,19 +327,21 @@ runTxBuildCmd
   -> TxBuildOutputOptions
   -> ExceptT ShelleyTxCmdError IO ()
 runTxBuildCmd
-  (AnyCardanoEra cEra) consensusModeParams@(AnyConsensusModeParams cModeParams) nid mScriptValidity mOverrideWits txins readOnlyRefIns
-  reqSigners txinsc mReturnColl mTotCollateral txouts changeAddr mValue mLowBound
-  mUpperBound certs wdrls metadataSchema scriptFiles metadataFiles mProtocolParamsFile mUpProp outputOptions = do
+    mNodeSocketPath (AnyCardanoEra cEra) consensusModeParams@(AnyConsensusModeParams cModeParams)
+    nid mScriptValidity mOverrideWits txins readOnlyRefIns
+    reqSigners txinsc mReturnColl mTotCollateral txouts changeAddr mValue mLowBound
+    mUpperBound certs wdrls metadataSchema scriptFiles metadataFiles mProtocolParamsFile mUpProp outputOptions = do
   -- The user can specify an era prior to the era that the node is currently in.
   -- We cannot use the user specified era to construct a query against a node because it may differ
   -- from the node's era and this will result in the 'QueryEraMismatch' failure.
 
-  SocketPath sockPath <- lift readEnvSocketPath & onLeft (left . ShelleyTxCmdSocketEnvError)
+  socketPath <- maybe (lift readEnvSocketPath) (pure . Right) mNodeSocketPath
+    & onLeft (left . ShelleyTxCmdSocketEnvError)
 
   let localNodeConnInfo = LocalNodeConnectInfo
                             { localConsensusModeParams = cModeParams
                             , localNodeNetworkId = nid
-                            , localNodeSocketPath = sockPath
+                            , localNodeSocketPath = unSocketPath socketPath
                             }
 
   AnyCardanoEra nodeEra <- lift (determineEra cModeParams localNodeConnInfo)
@@ -373,11 +376,12 @@ runTxBuildCmd
   let filteredTxinsc = Set.toList $ Set.fromList txinsc
 
   -- We need to construct the txBodycontent outside of runTxBuild
-  BalancedTxBody txBodycontent balancedTxBody _ _
-    <- runTxBuild cEra consensusModeParams nid mScriptValidity inputsAndMaybeScriptWits readOnlyRefIns filteredTxinsc
-                  mReturnCollateral mTotCollateral txOuts changeAddr valuesWithScriptWits mLowBound
-                  mUpperBound certsAndMaybeScriptWits withdrawalsAndMaybeScriptWits
-                  requiredSigners txAuxScripts txMetadata mpparams mProp mOverrideWits outputOptions
+  BalancedTxBody txBodycontent balancedTxBody _ _ <-
+    runTxBuild
+      socketPath cEra consensusModeParams nid mScriptValidity inputsAndMaybeScriptWits readOnlyRefIns filteredTxinsc
+      mReturnCollateral mTotCollateral txOuts changeAddr valuesWithScriptWits mLowBound
+      mUpperBound certsAndMaybeScriptWits withdrawalsAndMaybeScriptWits
+      requiredSigners txAuxScripts txMetadata mpparams mProp mOverrideWits outputOptions
 
   let allReferenceInputs = getAllReferenceInputs
                              inputsAndMaybeScriptWits
@@ -602,7 +606,8 @@ runTxBuildRaw era
       getIsCardanoEraConstraint era $ createAndValidateTransactionBody txBodyContent
 
 runTxBuild
-  :: CardanoEra era
+  :: SocketPath
+  -> CardanoEra era
   -> AnyConsensusModeParams
   -> NetworkId
   -> Maybe ScriptValidity
@@ -639,11 +644,12 @@ runTxBuild
   -> Maybe Word
   -> TxBuildOutputOptions
   -> ExceptT ShelleyTxCmdError IO (BalancedTxBody era)
-runTxBuild era (AnyConsensusModeParams cModeParams) networkId mScriptValidity
-           inputsAndMaybeScriptWits readOnlyRefIns txinsc mReturnCollateral mTotCollateral txouts
-           (TxOutChangeAddress changeAddr) valuesWithScriptWits mLowerBound mUpperBound
-           certsAndMaybeScriptWits withdrawals reqSigners txAuxScripts txMetadata mpparams
-           mUpdatePropF mOverrideWits outputOptions = do
+runTxBuild
+    (SocketPath sockPath) era (AnyConsensusModeParams cModeParams) networkId mScriptValidity
+    inputsAndMaybeScriptWits readOnlyRefIns txinsc mReturnCollateral mTotCollateral txouts
+    (TxOutChangeAddress changeAddr) valuesWithScriptWits mLowerBound mUpperBound
+    certsAndMaybeScriptWits withdrawals reqSigners txAuxScripts txMetadata mpparams
+    mUpdatePropF mOverrideWits outputOptions = do
 
   liftIO $ forM_ mpparams $ \_ ->
     printWarning "'--protocol-params-file' for 'transaction build' is deprecated"
@@ -680,8 +686,6 @@ runTxBuild era (AnyConsensusModeParams cModeParams) networkId mScriptValidity
       void $ pure (toEraInMode era CardanoMode)
         & onNothing (left (ShelleyTxCmdEraConsensusModeMismatchTxBalance outputOptions
                             (AnyConsensusMode CardanoMode) (AnyCardanoEra era)))
-
-      SocketPath sockPath <- lift readEnvSocketPath & onLeft (left . ShelleyTxCmdSocketEnvError)
 
       let allTxInputs = inputsThatRequireWitnessing ++ allReferenceInputs ++ txinsc
           localNodeConnInfo = LocalNodeConnectInfo
@@ -1106,12 +1110,14 @@ runTxSign txOrTxBody witSigningData mnw (TxFile outTxFile) = do
 
 
 runTxSubmit
-  :: AnyConsensusModeParams
+  :: Maybe SocketPath
+  -> AnyConsensusModeParams
   -> NetworkId
   -> FilePath
   -> ExceptT ShelleyTxCmdError IO ()
-runTxSubmit (AnyConsensusModeParams cModeParams) network txFilePath = do
-    SocketPath sockPath <- lift readEnvSocketPath & onLeft (left . ShelleyTxCmdSocketEnvError)
+runTxSubmit mNodeSocketPath (AnyConsensusModeParams cModeParams) network txFilePath = do
+    SocketPath sockPath <- maybe (lift readEnvSocketPath) (pure . Right) mNodeSocketPath
+      & onLeft (left . ShelleyTxCmdSocketEnvError)
 
     txFile <- liftIO $ fileOrPipe txFilePath
     InAnyCardanoEra era tx <- lift (readFileTx txFile) & onLeft (left . ShelleyTxCmdCddlError)